# docs/swagger.yaml
definitions:
  model.PostData:
    properties:
      avatar:
        type: string
      content:
        type: string
      create_timestamp:
        description: 创建时间戳（秒级）
        type: string
      id:
        type: integer
      title:
        type: string
      update_timestamp:
        description: 更新时间戳（秒级）
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  model.ResponseWrapper:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  model.TokenPairData:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  model.UploadAvatarData:
    properties:
      avatar_path:
        type: string
    type: object
  model.UserData:
    properties:
      avatar:
        type: string
      create_timestamp:
        description: 创建时间戳（秒级）
        type: string
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  model.UserLoginData:
    properties:
      token_pair_data:
        $ref: '#/definitions/model.TokenPairData'
      user_data:
        $ref: '#/definitions/model.UserData'
    type: object
info:
  contact:
    name: 微信公众号：一枫说码
  description: 本项目是一个前后端分离的多用户论坛项目，实现了用户注册、登录、发帖、评论、私信、点赞、关注、搜索、记录日志、敏感词过滤等功能。
  title: 微学堂
  version: "1.0"
paths:
  /post/create:
    post:
      consumes:
      - application/json
      description: 创建帖子
      parameters:
      - description: 标题
        in: query
        name: title
        required: true
        type: string
      - description: 正文
        in: query
        name: content
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code为0表示成功，否则为失败
          schema:
            allOf:
            - $ref: '#/definitions/model.ResponseWrapper'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/model.PostData'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建帖子
      tags:
      - 帖子
  /post/delete:
    post:
      consumes:
      - application/json
      description: 根据帖子ID删除帖子
      parameters:
      - description: 帖子ID
        in: query
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code为0表示成功，否则为失败
          schema:
            allOf:
            - $ref: '#/definitions/model.ResponseWrapper'
            - properties:
                code:
                  type: integer
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除帖子
      tags:
      - 帖子
  /post/get/{post_id}:
    get:
      consumes:
      - application/json
      description: 获取帖子
      parameters:
      - description: 帖子ID
        in: path
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code为0表示成功，否则为失败
          schema:
            allOf:
            - $ref: '#/definitions/model.ResponseWrapper'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/model.PostData'
              type: object
      summary: 获取帖子
      tags:
      - 帖子
  /post/list:
    get:
      consumes:
      - application/json
      description: 获取帖子列表，支持分页
      parameters:
      - default: 1
        description: 页码
        in: query
        minimum: 1
        name: page_num
        required: true
        type: integer
      - default: 10
        description: 每页数量
        in: query
        maximum: 50
        minimum: 1
        name: page_size
        required: true
        type: integer
      - description: 用户ID
        in: query
        minimum: 1
        name: user_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code为0表示成功，否则为失败
          schema:
            allOf:
            - $ref: '#/definitions/model.ResponseWrapper'
            - properties:
                code:
                  type: integer
                data:
                  items:
                    $ref: '#/definitions/model.PostData'
                  type: array
              type: object
      summary: 获取帖子列表
      tags:
      - 帖子
  /post/update:
    post:
      consumes:
      - application/json
      description: 更改帖子
      parameters:
      - description: 标题
        in: query
        name: title
        required: true
        type: string
      - description: 正文
        in: query
        name: content
        required: true
        type: string
      - description: 帖子ID
        in: query
        name: post_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code为0表示成功，否则为失败
          schema:
            allOf:
            - $ref: '#/definitions/model.ResponseWrapper'
            - properties:
                code:
                  type: integer
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更改帖子
      tags:
      - 帖子
  /user/get/{user_id}:
    get:
      consumes:
      - application/json
      description: 根据用户ID获取公开可见的用户信息
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: code为0表示成功，否则为失败
          schema:
            allOf:
            - $ref: '#/definitions/model.ResponseWrapper'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/model.UserData'
              type: object
      summary: 获取指定用户信息
      tags:
      - 用户
  /user/info:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户信息
      produces:
      - application/json
      responses:
        "200":
          description: code为0表示成功，否则为失败
          schema:
            allOf:
            - $ref: '#/definitions/model.ResponseWrapper'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/model.UserData'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取当前登录用户信息
      tags:
      - 用户
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户登录，返回token
      parameters:
      - description: 邮箱
        in: query
        name: email
        required: true
        type: string
      - description: 密码
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code为0表示成功，否则为失败
          schema:
            allOf:
            - $ref: '#/definitions/model.ResponseWrapper'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/model.UserLoginData'
              type: object
      summary: 用户登录
      tags:
      - 用户
  /user/refresh_token:
    post:
      consumes:
      - application/json
      description: 根据refresh_token获取新token
      parameters:
      - description: 刷新token
        in: query
        name: refresh_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code为0表示成功，否则为失败
          schema:
            allOf:
            - $ref: '#/definitions/model.ResponseWrapper'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/model.TokenPairData'
              type: object
      summary: 刷新token
      tags:
      - 用户
  /user/register:
    post:
      consumes:
      - application/json
      description: 创建用户
      parameters:
      - description: 用户名
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: password
        required: true
        type: string
      - description: 邮箱
        in: query
        name: email
        required: true
        type: string
      - description: 六位验证码
        in: query
        name: verify_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code为0表示成功，否则为失败
          schema:
            allOf:
            - $ref: '#/definitions/model.ResponseWrapper'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/model.UserData'
              type: object
      summary: 创建用户
      tags:
      - 用户
  /user/register_code:
    get:
      consumes:
      - application/json
      description: 获取注册邮箱验证码
      parameters:
      - description: 邮箱
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code为0表示成功，否则为失败
          schema:
            allOf:
            - $ref: '#/definitions/model.ResponseWrapper'
            - properties:
                code:
                  type: integer
              type: object
      summary: 获取注册邮箱验证码
      tags:
      - 用户
  /user/update/info:
    post:
      consumes:
      - application/json
      description: 更新用户信息
      parameters:
      - description: 用户名
        in: query
        name: username
        type: string
      - description: 头像
        in: query
        name: avatar
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code为0表示成功，否则为失败
          schema:
            allOf:
            - $ref: '#/definitions/model.ResponseWrapper'
            - properties:
                code:
                  type: integer
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新用户信息
      tags:
      - 用户
  /user/update/password:
    post:
      consumes:
      - application/json
      description: 更新密码
      parameters:
      - description: 旧密码
        in: query
        name: old_password
        required: true
        type: string
      - description: 新密码
        in: query
        name: new_password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code为0表示成功，否则为失败
          schema:
            allOf:
            - $ref: '#/definitions/model.ResponseWrapper'
            - properties:
                code:
                  type: integer
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新密码
      tags:
      - 用户
  /user/upload/avatar:
    post:
      consumes:
      - multipart/form-data
      description: 上传用户头像并返回存储路径
      parameters:
      - description: 头像文件
        in: formData
        name: avatar
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: code为0表示成功，否则为失败
          schema:
            allOf:
            - $ref: '#/definitions/model.ResponseWrapper'
            - properties:
                code:
                  type: integer
                data:
                  $ref: '#/definitions/model.UploadAvatarData'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 上传用户头像
      tags:
      - 用户
securityDefinitions:
  ApiKeyAuth:
    description: '使用格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"